Abstracted command processing to util
