+++ src/bot.js
@@ -1,11 +1,10 @@
 import './config'
 import fs from 'fs'
 import Discord from 'discord.js'
-import { argParser } from './utils/argParser'
-import dedent from 'dedent'
 import { cmd } from './utils/cmd'
 import { welcomeTemplate } from './data/templates'
 import Format from 'string-format'
+import { processCommand } from './utils/processCommand'
 
 const bot = new Discord.Client({
     presence: {
@@ -40,31 +39,7 @@ for (const file of commandFiles) {
 bot.uniqueCommands = [...new Set(bot.commands.array())]
 
 bot.on('message', (msg) => {
-    const escPrefix = process.env.PREFIX.split('')
-        .map((c) => `\\${c}`)
-        .join('')
-
-    const regexStr = `^${escPrefix}(\\w+)(?:\\ (\\S.*)|)$`
-
-    const commandSyntax = new RegExp(regexStr, 'i')
-    if (msg.author.bot) return
-    if (!msg.content.match(commandSyntax)) return
-
-    const [, command, argStr] = msg.content.match(commandSyntax)
-    const args = argParser(argStr)
-
-    // Show message when no command found
-    if (!bot.commands.has(command)) {
-        msg.channel.send(`Command '${command}' not found 😬`)
-        return
-    }
-
-    try {
-        bot.commands.get(command).execute(msg, args)
-    } catch (e) {
-        console.log(e)
-        msg.channel.send(`There was a problem trying to run '${command}'`)
-    }
+    processCommand(msg)
 })
 
 bot.on('guildMemberAdd', async (member) => {
